[tailor @ 821105ddeb18b03f4fba8bad8eb99b5c76db0ada] *For Daniel's multiline buy I moved the onsuccess logic to the callback so this will be executed only when the adapter is actually in multiline mode.
*I refactored a little the method that creates the fields in one shot checkout, now
they are being created correctly (calling set method where available, that fixed
the problem of title not showing as fields label), also they are being created
inside a frameset to split data in in billing, user and payment info.

*The Checkout process should be working, I spent a interesting ammount of time
refactoring it until I made it fit into our MakePaymentProcess. Still some of
the reaquired fields are missing so I cant not start creating tests for this.

*I Adapted Daniel's onsuccess method for the oneshoot checkout so I was able to call
MakePaymentProcess. Still needs a couple of fields and returning the value of the
validation.

Original author: Horacio Duran <hduran@except.com.ar>
Date: 2008-06-18 02:12:34-03:00